moved preview chart into datasetpicker
