+++ web/vue/src/components/backtester/result/backtestResultChart.vue
@@ -1,6 +1,5 @@
 <template lang='jade'>
 div
-  h3 Backtest Result Chart
   div#resultHighchart
 </template>
